+++ zcash_client_backend/src/data_api/testing/transparent.rs
@@ -1,21 +1,19 @@
 use crate::{
     data_api::{
-        testing::{AddressType, TestBuilder, TestState},
-        testing::{DataStoreFactory, ShieldedProtocol, TestCache},
+        testing::{
+            AddressType, DataStoreFactory, ShieldedProtocol, TestBuilder, TestCache, TestState,
+        },
         wallet::input_selection::GreedyInputSelector,
         Account as _, InputSource, WalletRead, WalletWrite,
     },
-    fees::{fixed, DustOutputPolicy},
+    fees::{standard, DustOutputPolicy, StandardFeeRule},
     wallet::WalletTransparentOutput,
 };
 use assert_matches::assert_matches;
 use sapling::zip32::ExtendedSpendingKey;
 use zcash_primitives::{
     block::BlockHash,
-    transaction::{
-        components::{amount::NonNegativeAmount, OutPoint, TxOut},
-        fees::fixed::FeeRule as FixedFeeRule,
-    },
+    transaction::components::{amount::NonNegativeAmount, OutPoint, TxOut},
 };
 
 pub fn put_received_transparent_utxo<DSF>(dsf: DSF)
@@ -198,8 +196,8 @@ where
 
     // Shield the output.
     let input_selector = GreedyInputSelector::new();
-    let change_strategy = fixed::SingleOutputChangeStrategy::new(
-        FixedFeeRule::non_standard(NonNegativeAmount::ZERO),
+    let change_strategy = standard::SingleOutputChangeStrategy::new(
+        StandardFeeRule::Zip317,
         None,
         ShieldedProtocol::Sapling,
         DustOutputPolicy::default(),